Pull out common helper functions into json_rpc_requests_helper
